+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added support for diagnostics & completions in [Cargo features](https://doc.rust-lang.org/cargo/reference/features.html)
+
 ## `0.4.1` - January 18th, 2025
 
 ### Added